fix: Rename Update to Modify
