+++ src/module/dy/fns/modify.js
